move duplicate checking into TypeCheck
This completes the effort to not touch the impl-items during
`Collect(Impl)`.
